+++ dart/lib/src/sentry_span_interface.dart
@@ -25,7 +25,7 @@ abstract class ISentrySpan {
   void removeData(String key);
 
   /// Sets span timestamp marking this span as finished.
-  Future<void> finish({SpanStatus? status}) async {}
+  Future<void> finish({SpanStatus? status, DateTime? endTimestamp}) async {}
 
   /// Gets span status.
   SpanStatus? get status;