+++ flutter/test/mocks.mocks.dart
@@ -86,8 +86,10 @@ class MockNoOpSentrySpan extends _i1.Mock implements _i2.NoOpSentrySpan {
       super.noSuchMethod(Invocation.setter(#status, status),
           returnValueForMissingStub: null);
   @override
-  _i7.Future<void> finish({_i3.SpanStatus? status}) =>
-      (super.noSuchMethod(Invocation.method(#finish, [], {#status: status}),
+  _i7.Future<void> finish({_i3.SpanStatus? status, DateTime? endTimestamp}) =>
+      (super.noSuchMethod(
+          Invocation.method(
+              #finish, [], {#status: status, #endTimestamp: endTimestamp}),
           returnValue: Future<void>.value(),
           returnValueForMissingStub: Future<void>.value()) as _i7.Future<void>);
   @override